FIX do not render empty groups UPDATE sorting of groups and properties
