+++ resources/lang/de/Widget.properties
@@ -602,5 +602,6 @@ dataFieldTextsMetaKeywords = "Meta-Keywords"
 ; Data Fields - Properties
 dataFieldProperties = "Eigenschaften"
 dataFieldPropertyGroupName = "Gruppenname"
+dataFieldPropertyGroupWithoutName = "Nicht gruppiert"
 dataFieldPropertyName = "Name"
 dataFieldPropertyValue = "Wert"
