+++ resources/lang/en/Widget.properties
@@ -602,5 +602,6 @@ dataFieldTextsMetaKeywords = "Meta keywords"
 ; Data Fields - Properties
 dataFieldProperties = "Properties"
 dataFieldPropertyGroupName = "Group name"
+dataFieldPropertyGroupWithoutName = "Ungrouped"
 dataFieldPropertyName = "Name"
 dataFieldPropertyValue = "Value"
